requirements.txt
freetype-py==2.2.0
glfw==2.1.0
numpy==1.20.2
Pillow==8.2.0
PyGLM==1.99.3
pylint==2.5.2
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5